requirements.txt
# Requirements
# Usage:
# pip install -r requirements.txt

# =================================BUILD REQUIREMENTS=================================
pyyaml
h5py==2.10.0
xlrd
openpyxl

# Web
beautifulsoup4
requests
selenium
pyperclip==1.8.0
PyAutoGUI==0.9.50

# GUI dev
pyqt5==5.15.4
qtawesome
pyinstaller==4.0

# Data analysis
numpy==1.19.5
pandas
missingno==0.4.2

# CV
opencv-python==4.4.0.42

# Visualization
matplotlib==3.1.3
pydot
graphviz==0.13.2

# Utils
tqdm==4.43.0
pytest==6.2.2
regex

# Flask Framework
bcrypt==3.1.4
blinker==1.4
certifi==2016.2.28
cffi==1.11.5
click==8.0.1
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Bootstrap-Flask==1.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.0
pycparser==2.18
six==1.11.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
WTForms==2.2.1
python-dotenv==0.18.0

# =================================DEV REQUIREMENTS=================================
coverage
flake8==3.8.3
autoflake==1.3.1
isort==5.5.3
black==19.10b0
jupyterlab
# for jupyter kernel
ipykernel

# =================================DOCS REQUIREMENTS=================================
mkdocs>=1.1
Pygments>=2.4
markdown>=3.2
pymdown-extensions>=7.0
mkdocs-material-extensions>=1.0
mkdocs-minify-plugin>=0.3
mkdocs-redirects>=1.0
